refactor: 자동차 이름, 시도할 횟수 유효성 검사하는 클래스, 각각 독립적인 함수로 변환
