+++ src/index.js
@@ -1,13 +1,13 @@
-import CheckValid from "./CheckValid.js";
 import Racing from "./Racing.js";
 import Render from "./Render.js";
 import User from "./User.js";
+import checkValidCarNames from "./validation/checkCarNames.js";
+import checkValidracingCount from "./validation/checkRacingCount.js";
 
 class RacingCarGame {
   constructor() {
     this.render = new Render();
     this.user = new User();
-    this.checkValid = new CheckValid();
     this.carNameForm = document.getElementById("car-names-form");
     this.racingCountForm = document.getElementById("racing-count-form");
   }
@@ -22,7 +22,7 @@ class RacingCarGame {
 
   handleSubmitCarName(event) {
     event.preventDefault();
-    const isValid = this.checkValid.carNames(this.user.getCarNames());
+    const isValid = checkValidCarNames(this.user.getCarNames());
     if (isValid) {
       this.render.showRacingCountForm();
       this.racingCountForm.addEventListener("submit", e =>
@@ -34,7 +34,7 @@ class RacingCarGame {
   handleSubmitRacingCount(event) {
     event.preventDefault();
     this.render.clearRacingResult();
-    const isValid = this.checkValid.racingCount(this.user.getRacingCount());
+    const isValid = checkValidracingCount(this.user.getRacingCount());
     if (isValid) {
       const racing = new Racing(
         this.user.getCarNames(),
